requirements.txt
fastapi
sqlmodel
alembic
minio
psycopg2
python-dotenv
pydantic-settings
